DC-1757: updating to make TwoWayMessageService a trait
