+++ app/uk/gov/hmrc/twowaymessage/services/TwoWayMessageService.scala
@@ -16,37 +16,41 @@
 
 package uk.gov.hmrc.twowaymessage.services
 
-import java.util.UUID.randomUUID
-
-import com.google.inject.Inject
+import com.google.inject.ImplementedBy
 import org.apache.commons.codec.binary.Base64
 import play.api.http.Status._
 import play.api.libs.json._
 import play.api.mvc.Result
 import play.api.mvc.Results._
-import play.twirl.api.Html
 import uk.gov.hmrc.domain.Nino
-import uk.gov.hmrc.gform.dms.{DmsHtmlSubmission, DmsMetadata}
-import uk.gov.hmrc.gform.gformbackend.GformConnector
+import uk.gov.hmrc.gform.dms.DmsMetadata
 import uk.gov.hmrc.http._
-import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
-import uk.gov.hmrc.twowaymessage.connectors.MessageConnector
-import uk.gov.hmrc.twowaymessage.enquiries.Enquiry
-import uk.gov.hmrc.twowaymessage.enquiries.Enquiry.EnquiryTemplate
 import uk.gov.hmrc.twowaymessage.model.{Error, _}
 import uk.gov.hmrc.twowaymessage.model.Error._
 import uk.gov.hmrc.twowaymessage.model.FormId.FormId
 import uk.gov.hmrc.twowaymessage.model.MessageType.MessageType
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.Future
 import scala.language.implicitConversions
-import scala.xml.Node
 
-object TwoWayMessageService {
+@ImplementedBy(classOf[TwoWayMessageServiceImpl])
+trait TwoWayMessageService {
 
   type ErrorFunction = (Int,String) => Result
 
   val errorResponse: ErrorFunction = (status: Int, message: String) => BadGateway(Json.toJson(Error(status, message)))
+
+  def getMessageMetaData(messageId: String)(implicit hc: HeaderCarrier): Future[MessageMetadata]
+
+  def post(queueId: String, nino: Nino, twoWayMessage: TwoWayMessage, dmsMetaData: DmsMetadata): Future[Result]
+
+  def postAdvisorReply(twoWayMessageReply: TwoWayMessageReply, replyTo: String)(implicit hc: HeaderCarrier): Future[Result]
+
+  def postCustomerReply(twoWayMessageReply: TwoWayMessageReply, replyTo: String)(implicit hc: HeaderCarrier): Future[Result]
+
+  def createDmsSubmission(html: String, response: HttpResponse, dmsMetaData: DmsMetadata): Future[Result]
+
+  def createHtmlMessage(messageId: String, nino: Nino, messageContent: String, subject: String): Future[Option[String]]
 
   def createJsonForMessage(refId: String, twoWayMessage: TwoWayMessage, nino: Nino, queueId: String): Message = {
     val recipient = Recipient(TaxIdentifier(nino.name, nino.value), twoWayMessage.contactDetails.email)
@@ -93,96 +97,6 @@ object TwoWayMessageService {
     case e: Upstream4xxResponse => errorResponse(e.upstreamResponseCode, e.message)
     case e: Upstream5xxResponse => errorResponse(e.upstreamResponseCode, e.message)
     case e: HttpException       => errorResponse(e.responseCode, e.message)
-  }
-
-}
-
-class TwoWayMessageService @Inject()(
-  messageConnector: MessageConnector,
-  gformConnector: GformConnector,
-  servicesConfig: ServicesConfig)(implicit ec: ExecutionContext) {
-
-  implicit val hc: HeaderCarrier = HeaderCarrier()
-
-  implicit def stringToXml(string: String): Seq[Node] = {
-    val xml = scala.xml.XML.loadString("<root>" + string + "</root>")
-    xml.child
-  }
-
-  def getMessageMetaData(messageId: String)(implicit hc: HeaderCarrier): Future[MessageMetadata] = messageConnector.getMessageMetadata(messageId)
-
-  def post(queueId: String, nino: Nino, twoWayMessage: TwoWayMessage, dmsMetaData: DmsMetadata): Future[Result] = {
-    val body = TwoWayMessageService.createJsonForMessage(randomUUID.toString, twoWayMessage, nino, queueId)
-    messageConnector.postMessage(body) flatMap { response =>
-      handleResponse(twoWayMessage.content, twoWayMessage.subject, response, dmsMetaData)
-    } recover TwoWayMessageService.handleError
-  }
-
-  def postAdvisorReply(twoWayMessageReply: TwoWayMessageReply, replyTo: String)(
-    implicit hc: HeaderCarrier): Future[Result] =
-    postReply(twoWayMessageReply, replyTo, MessageType.Advisor, FormId.Reply)
-
-  def postCustomerReply(twoWayMessageReply: TwoWayMessageReply, replyTo: String)(implicit hc: HeaderCarrier): Future[Result] =
-    (for {
-      metadata <- getMessageMetaData(replyTo)
-      queueId <- metadata.details.enquiryType
-          .fold[Future[String]](Future.failed(new Exception(s"Unable to get DMS queue id for $replyTo")))(Future.successful)
-      enquiryId <- Enquiry(queueId).fold[Future[EnquiryTemplate]](Future.failed(new Exception(s"Unknown $queueId")))(Future.successful)
-      dmsMetaData = DmsMetadata(enquiryId.dmsFormId, metadata.recipient.identifier.value, enquiryId.classificationType, enquiryId.businessArea)
-      body = TwoWayMessageService.createJsonForReply(randomUUID.toString, MessageType.Customer, FormId.Question, metadata, twoWayMessageReply, replyTo)
-      postMessageResponse <- messageConnector.postMessage(body)
-      dmsHandleResponse <- handleResponse(twoWayMessageReply.content, metadata.subject, postMessageResponse, dmsMetaData)
-    } yield dmsHandleResponse) recover TwoWayMessageService.handleError
-
-  def createDmsSubmission(html: String, response: HttpResponse, dmsMetaData: DmsMetadata): Future[Result] = {
-    val dmsSubmission = DmsHtmlSubmission(TwoWayMessageService.encodeToBase64String(html), dmsMetaData)
-    Future.successful(Created(Json.parse(response.body))).andThen {
-      case _ => gformConnector.submitToDmsViaGform(dmsSubmission)
-    }
-  }
-
-  def createHtmlMessage(messageId: String, nino: Nino, messageContent: String, subject: String): Future[Option[String]] = {
-    import XmlTransformService._
-    val frontendUrl: String = servicesConfig.getString("pdf-admin-prefix")
-    val url = s"$frontendUrl/message/$messageId/reply"
-    getMessageContent(messageId).flatMap {
-      case Some(content) =>
-        val htmlText = updateDatePara(stripH1(stripH2(content))).mkString
-        Future.successful(Some(uk.gov.hmrc.twowaymessage.views.html.two_way_message(url, nino.nino, subject, Html(htmlText)).body))
-      case None => Future.successful(None)
-    }
-  }
-
-  private def postReply(twoWayMessageReply: TwoWayMessageReply, replyTo: String, messageType: MessageType, formId: FormId)(
-    implicit hc: HeaderCarrier): Future[Result] =
-    (for {
-      metadata <- messageConnector.getMessageMetadata(replyTo)
-      body = TwoWayMessageService.createJsonForReply(randomUUID.toString, messageType, formId, metadata, twoWayMessageReply, replyTo)
-      resp <- messageConnector.postMessage(body)
-    } yield resp) map {
-      TwoWayMessageService.handleResponse
-    } recover TwoWayMessageService.handleError
-
-  private def handleResponse(content: String, subject: String, response: HttpResponse, dmsMetaData: DmsMetadata): Future[Result] =
-    response.status match {
-      case CREATED =>
-        response.json.validate[Identifier].asOpt match {
-          case Some(identifier) =>
-            createHtmlMessage(identifier.id, Nino(dmsMetaData.customerId), content, subject).flatMap {
-              case Some(html) => createDmsSubmission(html,response,dmsMetaData)
-              case _ => Future.successful(TwoWayMessageService.errorResponse(INTERNAL_SERVER_ERROR, "Failed to create HTML for DMS submission"))
-            }
-          case None => Future.successful(TwoWayMessageService.errorResponse(INTERNAL_SERVER_ERROR, "Failed to create enquiry reference"))
-        }
-      case _ => Future.successful(TwoWayMessageService.errorResponse(response.status, response.body))
-    }
-
-  def getMessageContent(messageId: String): Future[Option[String]] = {
-    messageConnector.getMessageContent(messageId).flatMap(response =>
-      TwoWayMessageService.getContent(response) match {
-        case Some(content) => Future successful Some(content)
-        case None => Future.successful(None)
-      })
   }
 
 }
